redesign the structure from catalog -menu_managemnt to ves_megamenu_menu in entity and model

+++ src/main/java/menuManegment/demo/menu/service/MenuManagementServiceImpl.java
@@ -30,11 +30,11 @@ public class MenuManagementServiceImpl extends AbstractService<MenuManagement,
     public List<MenuManagement> updateStatus(String value, List<MenuManagementModel> models) {
         if (value.equals("enable")) {
             for (MenuManagementModel model : models) {
-                model.setStatus(1);
+                model.setStatus((short) 1);
             }
         } else if (value.equals("disable")) {
             for (MenuManagementModel model : models) {
-                model.setStatus(0);
+                model.setStatus((short) 0);
             }
         }
